+++ scripts/CreateHistoricalYieldDataTable.sql
@@ -5,7 +5,7 @@ create table dbo.HistoricalYieldData
 	[Id] int identity primary key clustered,
 	[Name]			  varchar(50) not null,
 	[Year]			  int,
-	[Yield]			  float,
+	[Yield]			  decimal,
 	constraint UN_YieldAmount unique([Name],[Year])
 )
