+++ scripts/sp_CreateNewInvestment.sql
@@ -14,8 +14,8 @@ END
 GO
 
 CREATE PROCEDURE [dbo].[sp_CreateNewInvestment](@valuationDate as DATETIME, @investment as VARCHAR(50), @symbol as CHAR(20),
-				 @currency as CHAR(3), @scalingFactor as FLOAT, @shares as FLOAT,
-				 @totalCost as FLOAT, @closingPrice as FLOAT,
+				 @currency as CHAR(3), @scalingFactor as decimal, @shares as decimal,
+				 @totalCost as decimal, @closingPrice as decimal,
 				 @account as INT, @exchange as VARCHAR(10)) AS
 BEGIN
